Move all the ingests_common code into ingests_tracker; get it passing
