--- ingests_common/docker-compose.yml
@@ -1,4 +1,0 @@
-dynamodb:
-  image: peopleperhour/dynamodb
-  ports:
-    - "45678:8000"
