+++ ingests_tracker/docker-compose.yml
@@ -3,3 +3,7 @@ sqs:
   ports:
     - "9324:9324"
     - "4789:9324"
+dynamodb:
+  image: "peopleperhour/dynamodb"
+  ports:
+    - "45678:8000"
