+++ ingests_tracker/src/main/scala/uk/ac/wellcome/platform/storage/ingests_tracker/tracker/IngestError.scala
@@ -1,4 +1,5 @@
-package uk.ac.wellcome.platform.archive.common.ingests.tracker
+package uk.ac.wellcome.platform.storage.ingests_tracker.tracker
+
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagVersion
 import uk.ac.wellcome.platform.archive.common.ingests.models.{Callback, Ingest}
 import uk.ac.wellcome.storage.{
