+++ ingests_tracker/src/main/scala/uk/ac/wellcome/platform/storage/ingests_tracker/tracker/dynamo/DynamoIngestTracker.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.ingests.tracker.dynamo
+package uk.ac.wellcome.platform.storage.ingests_tracker.tracker.dynamo
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import grizzled.slf4j.Logging
@@ -6,7 +6,7 @@ import org.scanamo.auto._
 import org.scanamo.time.JavaTimeFormats._
 import uk.ac.wellcome.platform.archive.common.ingests.models.{Ingest, IngestID}
 import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID._
-import uk.ac.wellcome.platform.archive.common.ingests.tracker.IngestTracker
+import uk.ac.wellcome.platform.storage.ingests_tracker.tracker.IngestTracker
 import uk.ac.wellcome.storage.dynamo._
 import uk.ac.wellcome.storage.store.VersionedStore
 import uk.ac.wellcome.storage.store.dynamo.DynamoSingleVersionStore
