+++ ingests_tracker/src/main/scala/uk/ac/wellcome/platform/storage/ingests_tracker/tracker/memory/MemoryIngestTracker.scala
@@ -1,7 +1,7 @@
-package uk.ac.wellcome.platform.archive.common.ingests.tracker.memory
+package uk.ac.wellcome.platform.storage.ingests_tracker.tracker.memory
 
 import uk.ac.wellcome.platform.archive.common.ingests.models.{Ingest, IngestID}
-import uk.ac.wellcome.platform.archive.common.ingests.tracker.IngestTracker
+import uk.ac.wellcome.platform.storage.ingests_tracker.tracker.IngestTracker
 import uk.ac.wellcome.storage.store.memory.MemoryVersionedStore
 
 class MemoryIngestTracker(
